Validation API

+++ Blazor.Domain/Entities/Venda.cs
@@ -24,13 +24,13 @@ namespace Domain.Entities
 
         public decimal ValorCompra { get; set; }
 
-        public double PercentualImposto { get; set; }
+        public decimal PercentualImposto { get; set; }
 
         public decimal ValorFrete1 { get; set; }
 
         public decimal ValorFrete2 { get; set; }
 
-        public double PercentualCustoOperacional { get; set; }
+        public decimal PercentualCustoOperacional { get; set; }
 
         public decimal ValorOutrasDespesas { get; set; }
 
@@ -40,7 +40,7 @@ namespace Domain.Entities
 
         public decimal ValorComissao { get; set; }
 
-        public double PercentualLucro { get; set; }
+        public decimal PercentualLucro { get; set; }
 
         public decimal ValorLucro { get; set; }
